Unify duplicated group attributes handling
